user-assistant/user_assistant_backend/requirements.txt
django
django-rest-framework
langchain<1.0.0
langchain-openai
langchain-community
redis
langchain-community
django-redis